usbipd builds usb device list on demand.
In linux implementation, usbipd maintains usb device list in memory.
But it does not seem to be effective because the list is refreshed at
every usbip request. Even worse, more elaborated approach is required
to prevent the race between listing and binding.
For now, a simple on-demand device listing and binding will be used.

Also common codes between usbipd and usbip will go into a userspace
library. For this effort, this commit has devno related common stuff
which goes into the library.