Fix BSOD caused when a partially sent urbr is unlinked
This commit will resolve #14.
List entries of a newly created urbr had not been initialized.
They was expected to be setup when interted into a pending or
a sent list. However, there's a missing point. For safety,
list entries of a new urbr will be initialized at creation
time. In addition, free_urbr() is added for sanity check of a
being destroyed urbr.