Allow a vhci iso pipe to be reset
An old vhci had not allowed an isochronous pipe to be reset.
But there was no reason to disallow an iso or a control pipe. And a
certain USB camera could not be detached without allowing an iso
pipe to be reset. This commit allows all pipes to be reset except a
control pipe. Such an restriction might be removed later.