introduce a new vhci arhitecture supporting virtual HCI and HUB
A new vhci driver devides its single device into a host contoller and
a virtual root hub. This will help a USB application to regard
a usbip-win as a real USB device. However, current implementation has
still problems and is unstable. Changes are as follows:
- 4 device types
  - vhci: fdo for virtual host controller
  - hpdo: pdo for virtual root hub
  - vhub: fdo for virtual root hub
  - vpdo: pdo for virtual USB device
- new vhci hardware id, device names

This commit has too many features and code beautifications. Sorry for
the complex commit.