Let vhci and vhub use USB as a device setup class
A custom device setup class(USBIPWin) prevents some USB apps such as
SimpleWMIView to detect a usbip-win virtual host controller and hub.
A new non-pnp root device and cpdo(controller pdo) are introduced
for devices of a vhci driver to be recoginized as USB device setup
class. Now, there are 6 vdev types in a vhci driver and 2 driver inf.
This fact causes that many pnp routines are reorganized and an usbip
install routine is changed a lot.